FIX/FEAT: Fix some error from printf and add functions

+++ ft_putnb.c
@@ -6,16 +6,16 @@
 /*   By: cda-fons <cda-fons@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/03 19:11:54 by cda-fons          #+#    #+#             */
-/*   Updated: 2024/05/04 17:55:41 by cda-fons         ###   ########.fr       */
+/*   Updated: 2024/05/05 16:00:13 by cda-fons         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_putnb(int nbr)
+int	ft_putnb(long nbr)
 {
 	int	i;
-	int nb;
+	long nb;
 	
 	i = 1;
 	nb = nbr;
@@ -29,7 +29,7 @@ int	ft_putnb(int nbr)
 		nb = nb / 10;
 		i++;
 	}
-	ft_putnbr_fd(nbr, 1);
+	ft_putnbr_fd((int)nbr, 1);
 	if (nbr == -2147483648)
 		return (11);
 	return (i);
